* Deshabilito la posibilidad de eliminar media mientras tenga headlines asociados * Ajustes al HeadlinesFeedList, orden por fecha descendnete, agrego los divs para mostrar mensajes y resultados, y ajustes de diseño * Fix en DocumentsUploadAction agrego parametros para el forward

+++ trunk/GCABA/mdp/WEB-INF/classes/modules/medias/actions/MediasDoDeleteAction.php
@@ -13,7 +13,10 @@ class MediasDoDeleteAction extends BaseAction {
 		if (isset($_POST["id"])) {
 			$media = MediaQuery::create()->findPK($_POST["id"]);
 			if (!empty($media)) {
-				$media->delete();
+				if ($media->countHeadlines() > 0)
+					return $mapping->findForwardConfig('failure');
+				else
+					$media->delete();
 				if ($media->isDeleted()) {
 					if (mb_strlen($media->getName()) > 120)
 						$cont = " ... ";
@@ -23,6 +26,8 @@ class MediasDoDeleteAction extends BaseAction {
 
 					return $mapping->findForwardConfig('success');
 				}
+				else
+					return $mapping->findForwardConfig('failure');
 			}
 		}
 		return $mapping->findForwardConfig('failure');
